Adds String#constantize from ActiveSupport
I don't need all of ActiveSupport for one method, although this gem
will mostly be used inside a Rails app. This removes the dependency
on Rails, and allows it to be used outside of a Rails app.

I also updated the LICENSE.txt file, giving attribution to both
ActiveSupport and Spice.